requirements.txt
google-genai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.0.0
z3-solver>=4.12.0
